+++ backend/services/embeddings_service.py
